# docs/openapi-leanschool.yaml
openapi: 3.0.3
info:
  title: leanschool API
  description: CRUD data layer for school management (receipts, locations, persons,
    classes, etc.).
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: local
paths:
...
  /grades:
    get:
      summary: List grades
      operationId: listGrades
      responses:
        '200':
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grade'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a grade
      operationId: createGrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeRequest'
      responses:
        '201':
          description: Created Grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /grades/{id}:
    get:
      summary: Get a Grade by ID
      operationId: getGrade
      parameters:
      - name: id
        in: path
        required: true
        description: Grade ID
        schema:
          type: string
      responses:
        '200':
          description: Grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '404':
          description: Not found
    put:
      summary: Update a Grade
      operationId: updateGrade
      parameters:
      - name: id
        in: path
        required: true
        description: Grade ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeRequest'
      responses:
        '200':
          description: Updated Grade
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a Grade
      operationId: deleteGrade
      parameters:
      - name: id
        in: path
        required: true
        description: Grade ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /locations:
    get:
      summary: List locations
      operationId: listLocations
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a location
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '201':
          description: Created Location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /locations/{id}:
    get:
      summary: Get a Location by ID
      operationId: getLocation
      parameters:
      - name: id
        in: path
        required: true
        description: Location ID
        schema:
          type: string
      responses:
        '200':
          description: Location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: Not found
    put:
      summary: Update a Location
      operationId: updateLocation
      parameters:
      - name: id
        in: path
        required: true
        description: Location ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '200':
          description: Updated Location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a Location
      operationId: deleteLocation
      parameters:
      - name: id
        in: path
        required: true
        description: Location ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /postal-codes:
    get:
      summary: List postal codes
      operationId: listPostalCodes
      responses:
        '200':
          description: List of postal codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostalCode'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a postal code
      operationId: createPostalCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostalCode'
      responses:
        '201':
          description: Created PostalCode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalCode'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /postal-codes/{id}:
    get:
      summary: Get a PostalCode by ID
      operationId: getPostalCode
      parameters:
      - name: id
        in: path
        required: true
        description: PostalCode ID
        schema:
          type: string
      responses:
        '200':
          description: PostalCode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalCode'
        '404':
          description: Not found
    put:
      summary: Update a PostalCode
      operationId: updatePostalCode
      parameters:
      - name: id
        in: path
        required: true
        description: PostalCode ID
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostalCode'
      responses:
        '200':
          description: Updated PostalCode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalCode'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Conflict — optimistic lock
    delete:
      summary: Delete a PostalCode
      operationId: deletePostalCode
      parameters:
      - name: id
        in: path
        required: true
        description: PostalCode ID
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
  schemas:
    Address:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: string
          example: 1a
        postalCode:
          $ref: '#/components/schemas/PostalCode'
        street:
          type: string
          example: Bahnhofstrasse
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    Building:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        location:
          $ref: '#/components/schemas/Location'
        name:
          type: string
          example: Main Building
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - version
...
    EntityRef:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      required:
      - id
    Exam:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        schoolClass:
          $ref: '#/components/schemas/SchoolClass'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    Grade:
      type: object
      properties:
        exam:
          $ref: '#/components/schemas/Exam'
        grade:
          type: number
          format: double
          example: 5.5
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        student:
          $ref: '#/components/schemas/Student'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    GradeRequest:
      type: object
      properties:
        exam:
          $ref: '#/components/schemas/EntityRef'
        grade:
          type: number
          format: double
          example: 5.5
        student:
          $ref: '#/components/schemas/EntityRef'
        version:
          type: integer
          example: 0
      required:
      - version
    Guardian:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        dateOfBirth:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - version
...
    Location:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        lat:
          type: number
          format: double
          example: 47.3769
        lon:
          type: number
          format: double
          example: 8.5417
        version:
          type: integer
          example: 0
      required:
      - id
      - version
...
    PostalCode:
      type: object
      properties:
        city:
          type: string
          example: Zurich
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        number:
          type: integer
          example: 8001
        version:
          type: integer
          example: 0
      required:
      - id
      - city
      - version
      - number
...
    Room:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/Building'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Room 101
        version:
          type: integer
          example: 0
        roomType:
          $ref: '#/components/schemas/RoomType'
      required:
      - id
      - name
      - version
...
    SchoolClass:
      type: object
      properties:
        classroom:
          $ref: '#/components/schemas/Room'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Class 5A
        schoolYear:
          $ref: '#/components/schemas/SchoolYear'
        shortcut:
          type: string
          example: 5A
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        teachers:
          type: array
          items:
            $ref: '#/components/schemas/Teacher'
        version:
          type: integer
          example: 0
      required:
      - id
      - version
      - name
      - shortcut
      - teachers
      - students
...
    SchoolYear:
      type: object
      properties:
        from:
          type: string
          format: date-time
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        to:
          type: string
          format: date-time
        version:
          type: integer
          example: 0
      required:
      - id
      - version
    Student:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        currentSchoolClass:
          $ref: '#/components/schemas/SchoolClass'
        dateOfBirth:
          type: string
          format: date-time
        grades:
          type: array
          items:
            $ref: '#/components/schemas/Grade'
        guardians:
          type: array
          items:
            $ref: '#/components/schemas/Guardian'
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Doe
        pastSchoolClasses:
          type: array
          items:
            $ref: '#/components/schemas/SchoolClass'
        prename:
          type: string
          example: Jane
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - guardians
      - version
      - grades
      - pastSchoolClasses
...
    Teacher:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        atSchoolSince:
          type: string
          format: date-time
        atSchoolUntil:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        iban:
          type: string
          example: CH93-0076-2011-6238-5295-7
        id:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Smith
        prename:
          type: string
          example: John
        sub:
          type: string
          example: keycloak-subject-uuid
        version:
          type: integer
          example: 0
      required:
      - id
      - name
      - prename
      - atSchoolSince
      - version
...
    RoomType:
      type: string
      enum:
      - ClassRoom
      - Kindergarden
      - Music
      - Video
      - Library
      - TeacherRoom
      - ComputerRoom
      - FacilityManagementRoom
      - Theater
      - MeetingRoom
      - ScienceRoom
      - SchoolManagementRoom
      - SchoolSecretariatRoom
...
  securitySchemes:
    keycloak:
      type: openIdConnect
      description: Keycloak OpenID Connect — JWT bearer token with realm_access roles
        (leanschool_read, leanschool_write)
      openIdConnectUrl: http://localhost:8180/realms/leanschool/.well-known/openid-configuration
security:
- keycloak: []